+++ lib/taro/types/shared/derived_types.rb
@@ -35,9 +35,11 @@ module Taro::Types::Shared::DerivedTypes
 
     root.define_singleton_method(method_name) do
       derived_types[type] ||= begin
+        name || raise(Taro::ArgumentError, 'Cannot derive from anonymous type')
+
         type_class = Taro::Types::Coercion.call(type:)
         new_type = Class.new(type_class)
-        new_type.define_name("#{self.name}.#{method_name}")
+        new_type.define_name("#{name}.#{method_name}")
         new_type.derive_from(self)
         new_type
       end